+++ CHANGELOG.md
@@ -5,6 +5,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Bumped runtimeVersion to `2024.05.28`.
  * Upgraded stable Dart analysis SDK to `3.4.1`
  * Upgraded stable Flutter analysis SDK to `3.22.1`.
+ * Note: backfilling new GitHub publishing fields.
 
 ## `20240523t085100-all`
  * Bumped runtimeVersion to `2024.05.22`.